requirements.txt
python-dotenv==1.0.1
aiogram==3.13.1
openai==1.51.2